Iterations of mixins are set in home route. Requires refactoring prior to Socket.io implementation.
Iterations of mixins are set in home route. Requires refactoring prior
to Socket.io implementation.

—————————

All mixin tests have been moved out of the sandbox route and are
working properly in the home route. Error route works. Refactoring to
be done to eliminate some extraneous code, especially that which stores
dynamic data in res.locals. Static data, like followers count and the
authenticated user’s avatar will remain is res.locals.

Next step after refactoring is to hook up socket.io and to create the
post request for the home route to enable creation of a new Tweet
object to be appended to the Timeline data on submit of tweet input
form.