Prepending data with web sockets into the app's "timeline" area works.
Prepending data with web sockets into the app's "timeline" area works.

———————

Socket.io is taking data from the client, passing it to the server, and
receiving the data on the client by prepending the data to the user
“timeline” area.

What remains to be done is to include a Twit POST request in the
io.sockets.emit( ) event on the server so that all of a tweet object’s
data is properly passed to the client and so that a valid tweet is
created from the application.